Move config stuff to config.rs
